API producing artifacts of all pre-execution files

+++ gpt_engineer/API/api.py
@@ -138,18 +138,21 @@ async def step_handler(step: Step) -> Step:
 
     # check if new files have been created and make artifacts for those (CURRENTLY ONLY CONSIDERS TOP LEVEL DIRECTORY AND WILL MAKE FALSE OVERWRITES IF THERE ARE MULTIPLE FILES WITH THE SAME NAME IN THE TREE).
     artifacts = await Agent.db.list_artifacts(step.task_id)
-    existing_artifacts = {artifact.file_name for artifact in artifacts}
+    existing_artifacts = {
+        os.path.join(artifact.relative_path, artifact.file_name) for artifact in artifacts
+    }
 
     # HACK SOLVING A TEMPORARY PROBLEM: CURRENTLY GPT-ENGINEER WRITES AND EXECUTES CODE IN A SUBDIR CALLED WORKSPACE BY DEFAULT, WHICH NOW IS A DIRECTORY INSIDE 'workspace_dir'. FOR CORRECT REPORTING, WE MUST COPY ALL FILES TO 'workspace_dir
 
     gpte_workspace_path = Path(os.path.join(workspace_dir, "workspace"))
-
-    for item in gpte_workspace_path.iterdir():
-        if item.is_dir():
-            (Path(workspace_dir) / item.name).mkdir(parents=True, exist_ok=True)
-            shutil.copytree(item, Path(workspace_dir) / item.name, dirs_exist_ok=True)
-        else:
-            shutil.copy2(item, workspace_dir)
+    if os.path.exists(gpte_workspace_path):
+        for item in gpte_workspace_path.iterdir():
+            if item.is_dir():
+                (Path(workspace_dir) / item.name).mkdir(parents=True, exist_ok=True)
+                shutil.copytree(item, Path(workspace_dir) / item.name, dirs_exist_ok=True)
+            else:
+                shutil.copy2(item, workspace_dir)
+        shutil.rmtree(gpte_workspace_path)
 
     # create artifacts, enabling agbenchmark to know about the existence of the files. LAST TIME I CHECKED, ANY NON-EMPTY RELATIVE PATHS GAVE RUNTIME ERRORS IN agbenchmark
     for item in os.listdir(workspace_dir):
@@ -162,6 +165,43 @@ async def step_handler(step: Step) -> Step:
                     relative_path="",
                     file_name=item,
                 )
+    # additionally, if the file pre-execution files exists, add the paths inside of it too
+    if os.path.exists(os.path.join(workspace_dir, "pre-execution-files.txt")):
+        with open(os.path.join(workspace_dir, "pre-execution-files.txt"), "r") as file:
+            # Iterate over each line in the file
+            for line in file:
+                line = line.strip()
+                if line not in existing_artifacts:
+                    directory, filename = os.path.split(line)
+                    existing_artifacts.add(line)
+                    await Agent.db.create_artifact(
+                        task_id=step.task_id,
+                        relative_path=directory,
+                        file_name=filename,
+                    )
+
+    # path_black_list = list()
+    # for dirpath, dirnames, filenames in os.walk(workspace_dir):
+    #     if "pyvenv.cfg" in filenames or "pip-selfcheck.json" in filenames:
+    #         path_black_list.append(dirpath)
+    #         continue
+    #     bools = [dirpath in path_name for path_name in path_black_list]
+    #     if any(bools):
+    #         continue
+    #     for filename in filenames:
+    #         full_path = os.path.join(dirpath, filename)
+    #
+    #         if (not full_path in existing_artifacts):
+    #             existing_artifacts.add(full_path)
+    #             if os.path.isfile(full_path):
+    #                 rel_path = os.path.relpath(dirpath, workspace_dir)
+    #                 # BENCHMARK_TEMP = "/home/axel/Software/Auto-GPT/benchmark/agbenchmark_config/temp_folder/"
+    #                 # Path(os.path.join(BENCHMARK_TEMP, rel_path)).mkdir(exist_ok=True, parents=True)
+    #                 await Agent.db.create_artifact(
+    #                     task_id=step.task_id,
+    #                     relative_path=str(rel_path),
+    #                     file_name=filename,
+    #                     )
 
     return step
